fix addRoutes
